add policy

+++ app/Http/Controllers/MediaPlanController.php
@@ -6,16 +6,22 @@ use App\Http\Requests\MediaPlanRequest;
 use App\Http\Resources\MediaPlan\MediaPlanCollection;
 use App\Http\Resources\MediaPlan\MediaPlanResource;
 use App\Models\MediaPlan;
+use Illuminate\Support\Facades\Gate;
 
 class MediaPlanController extends Controller
 {
-    public function index(): \Illuminate\Http\Resources\Json\AnonymousResourceCollection
+    public function index(): \Illuminate\Http\JsonResponse|\Illuminate\Http\Resources\Json\AnonymousResourceCollection
     {
+        // Policy check
+        Gate::authorize('viewAny', MediaPlan::class);
+
         return MediaPlanCollection::collection(MediaPlan::all());
     }
 
     public function store(MediaPlanRequest $request): \Illuminate\Http\Response|MediaPlanResource
     {
+        Gate::authorize('create', MediaPlan::class);
+
         $validatedData = $request->validated();
 
         $mediaPlan = MediaPlan::create($validatedData);
@@ -25,11 +31,15 @@ class MediaPlanController extends Controller
 
     public function show(MediaPlan $mediaPlan): MediaPlanResource
     {
+        Gate::authorize('view', $mediaPlan);
+
         return new MediaPlanResource($mediaPlan);
     }
 
     public function update(MediaPlanRequest $request, MediaPlan $mediaPlan): \Illuminate\Http\Response|MediaPlanResource
     {
+        Gate::authorize('update', $mediaPlan);
+
         $validatedData = $request->validated();
 
         $mediaPlan->update($validatedData);
@@ -39,6 +49,8 @@ class MediaPlanController extends Controller
 
     public function destroy(MediaPlan $mediaPlan): \Illuminate\Http\JsonResponse
     {
+        Gate::authorize('delete', $mediaPlan);
+
         $deleted = $mediaPlan->delete();
 
         if ($deleted) {
